requirements.txt
Django==2.0.6
PyMySQL==0.9.3
jieba==0.39
requests==2.21.0
Pillow==5.4.1
djangorestframework==3.9.3
